- fixed cyclic references
